Decouple camera backtrack and player respawn
