+++ Assets/Scripts/DesirePaths/GameManager.cs
@@ -192,18 +192,6 @@ namespace DesirePaths
         /// <param name="position"></param>
         void RespawnPlayer(bool safe, Vector3 position)
         {
-            //du coup dans ta note rajoute le fait que je dois deposer les viscères avant de faire le respawn, et le deposite cadaver - si on veut faire une animation qui backtrack tout notre chemin
-            /*_playerSpawner.OnPlayerRespawnComplete.AddListener(delegate
-            {
-                _playerSpawner.OnPlayerRespawnComplete.RemoveAllListeners();
-                if (!safe)
-                {
-                    //Deposite cadaver if not on safe space
-                    _cadaverManager.DepositCadaverOnPosition(position);
-                } //We place only when we're sure camera is not looking, so when resuscitate is call                
-                _playerHealth.Resuscitate();
-            });*/
-
             _uiManager.HideNShow(); //Fade
             if (!safe)
             {
