+++ Assets/Scripts/DesirePaths/UX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc76c32662caf63458cacc1f2b8f10ef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
