+++ Assets/Scripts/DesirePaths/UX/CameraBacktrack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4b973e85a2515f94c976b1dd20f2bad9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
